# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole =
  | (typeof UserProfileRole)[keyof typeof UserProfileRole]
  | null;

export const UserProfileRole = {
  admin: "admin",
  member: "member",
} as const;

export interface UserProfile {
  id: string;
  email: string;
  firstName?: string | null;
  lastName?: string | null;
  imageUrl?: string | null;
  role?: UserProfileRole;
}

export type ProjectMyRole = (typeof ProjectMyRole)[keyof typeof ProjectMyRole];

export const ProjectMyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface Project {
  id: number;
  name: string;
  description?: string | null;
  ownerId: string;
  memberCount: number;
  taskCount: number;
  completedTaskCount: number;
  myRole: ProjectMyRole;
  createdAt: string;
  updatedAt: string;
}

export type ProjectMemberRole =
  (typeof ProjectMemberRole)[keyof typeof ProjectMemberRole];

export const ProjectMemberRole = {
  admin: "admin",
  member: "member",
} as const;

export interface ProjectMember {
  userId: string;
  projectId: number;
  role: ProjectMemberRole;
  joinedAt: string;
  user: UserProfile;
}

export type ProjectDetail = Project & {
  members: ProjectMember[];
};

export interface CreateProjectBody {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  description?: string | null;
}

export interface UpdateProjectBody {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  description?: string | null;
}

export type AddMemberBodyRole =
  (typeof AddMemberBodyRole)[keyof typeof AddMemberBodyRole];

export const AddMemberBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface AddMemberBody {
  userId: string;
  role: AddMemberBodyRole;
}

export type UpdateMemberBodyRole =
  (typeof UpdateMemberBodyRole)[keyof typeof UpdateMemberBodyRole];

export const UpdateMemberBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface UpdateMemberBody {
  role: UpdateMemberBodyRole;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  projectId: number;
  title: string;
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  assigneeId?: string | null;
  assignee?: UserProfile | null;
  creatorId: string;
  dueDate?: string | null;
  projectName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyStatus =
  (typeof CreateTaskBodyStatus)[keyof typeof CreateTaskBodyStatus];

export const CreateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  description?: string | null;
  status?: CreateTaskBodyStatus;
  priority?: CreateTaskBodyPriority;
  assigneeId?: string | null;
  dueDate?: string | null;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title?: string;
  description?: string | null;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  assigneeId?: string | null;
  dueDate?: string | null;
}

export type DashboardSummaryTasksByPriority = {
  high: number;
  medium: number;
  low: number;
};

export interface DashboardSummary {
  totalProjects: number;
  totalTasks: number;
  completedTasks: number;
  inProgressTasks: number;
  todoTasks: number;
  overdueTasks: number;
  myAssignedTasks: number;
  tasksByPriority: DashboardSummaryTasksByPriority;
}

export type ListProjectTasksParams = {
  status?: ListProjectTasksStatus;
  assigneeId?: string;
};

export type ListProjectTasksStatus =
  (typeof ListProjectTasksStatus)[keyof typeof ListProjectTasksStatus];

export const ListProjectTasksStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;