# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Task Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/members:
    get:
      operationId: listProjectMembers
      tags: [members]
      summary: List project members
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
    post:
      operationId: addProjectMember
      tags: [members]
      summary: Add a member to project (Admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberBody"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
  /projects/{projectId}/members/{userId}:
    put:
      operationId: updateProjectMember
      tags: [members]
      summary: Update member role (Admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberBody"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
    delete:
      operationId: removeProjectMember
      tags: [members]
      summary: Remove a member from project (Admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary for current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/my-tasks:
    get:
      operationId: getMyTasks
      tags: [dashboard]
      summary: Get all tasks assigned to current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Tasks assigned to me
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /dashboard/overdue:
    get:
      operationId: getOverdueTasks
      tags: [dashboard]
      summary: Get overdue tasks across all user projects
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Overdue tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, member]
          nullable: true
      required:
        - id
        - email
...
    ProjectMember:
      type: object
      properties:
        userId:
          type: string
        projectId:
          type: integer
        role:
          type: string
          enum: [admin, member]
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - userId
        - projectId
        - role
        - joinedAt
        - user
    AddMemberBody:
      type: object
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [admin, member]
      required:
        - userId
        - role
    UpdateMemberBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, member]
      required:
        - role
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high]
        assigneeId:
          type: string
          nullable: true
        assignee:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
        creatorId:
          type: string
        dueDate:
          type: string
          format: date-time
          nullable: true
        projectName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - status
        - priority
        - creatorId
        - createdAt
        - updatedAt
...
    DashboardSummary:
      type: object
      properties:
        totalProjects:
          type: integer
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        inProgressTasks:
          type: integer
        todoTasks:
          type: integer
        overdueTasks:
          type: integer
        myAssignedTasks:
          type: integer
        tasksByPriority:
          type: object
          properties:
            high:
              type: integer
            medium:
              type: integer
            low:
              type: integer
          required:
            - high
            - medium
            - low
      required:
        - totalProjects
        - totalTasks
        - completedTasks
        - inProgressTasks
        - todoTasks
        - overdueTasks
        - myAssignedTasks
        - tasksByPriority